server/api/requirements.txt
psycopg2-binary
clickhouse_driver

fastapi
uvicorn[standard]

pika
aio_pika

python-dotenv

asyncio